[issue#7] remove warnings in parallel simulation
Still big warnings in pic simulation

+++ src/simulations/parallel/bsl_dk_3d1v_polar/sll_m_sim_bsl_dk_3d1v_polar.F90
@@ -1986,8 +1986,8 @@ contains
       SLL_ALLOCATE(x2(nnodes_x1,nnodes_x2), error)
       do j = 1,nnodes_x2
         do i = 1,nnodes_x1
-          r       = rmin+real(i-1,f32)*dr
-          theta   = real(j-1,f32)*dtheta
+          r       = rmin+real(i-1,f64)*dr
+          theta   = real(j-1,f64)*dtheta
           x1(i,j) = r*cos(theta)
           x2(i,j) = r*sin(theta)
         end do
